requirements.txt
pandas
pygoogle-image-download
Pillow
python-dotenv
